requirements.txt
affine==2.4.0
attrs==24.3.0
certifi==2024.12.14
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
numpy==2.2.1
pyparsing==3.2.1
pystac==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rasterio==1.4.3
shapely==2.0.6
six==1.17.0
